delete user

+++ ssa/api/users.py
@@ -3,7 +3,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from typing import List
 
 from ssa import models, schemas, crud
-from ssa.dependencies import get_session
+from ssa.dependencies import get_session, verify_user_by_email, verify_user
 
 
 router = APIRouter(
@@ -11,7 +11,8 @@ router = APIRouter(
     tags=["users"],
     responses = {
         404 : {"description": "Not found"}, 
-        400 : {"description": "Bad request"}
+        400 : {"description": "Bad request"},
+        409 : {"description": "Conflict"},
     }
 )
 
@@ -38,7 +39,13 @@ async def get_user(user_id: int, db: AsyncSession = Depends(get_session)):
 
 @router.delete("/")
 async def delete_user(user: schemas.UserDelete, db: AsyncSession = Depends(get_session)):
-    res = await crud.delete_user(db, user)
-    if not res:
+    verified = await verify_user_by_email(db, email=user.email, password=user.password)
+    if not verified:
         raise HTTPException(status_code=404, detail="Password not correct or user not registered.")
+
+    algs_db = await crud.get_algorithms_by_author_email(db, email=user.email)
+    if algs_db:
+        raise HTTPException(status_code=409, detail="Cannot delete author of existing algorithms.") 
+    
+    res = await crud.delete_user(db, user)
     return {"ok" : res}
